refactor: consolidate TTS functionality by merging apiTTS into useTTS and streamline audio playback methods
